+++ services/client/src/Common/Components/Boxes/TuningTick.tsx
@@ -6,7 +6,7 @@ export const TuningTick = styled(Box)(({ theme }) => ({
   height: '20px',
   backgroundColor: theme.palette.secondary.dark,
   cursor: 'pointer',
-  transition: 'backgroundColor 0.2s ease-in-out',
+  transition: 'all 0.2s ease-in-out',
   boxShadow: '0 4px 8px rgba(0,0,0,.6)',
   '&:hover': {
     backgroundColor: theme.palette.secondary.main,
